chore: implement e2e test mvp
Foundation to consistently verify against `terraform` cli proper
