+++ preview_test.go
@@ -2,134 +2,19 @@ package preview_test
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"io/fs"
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/zclconf/go-cty/cty"
 	"github.com/zclconf/go-cty/cty/gocty"
-	"golang.org/x/exp/slices"
 
 	"github.com/coder/preview"
-	"github.com/coder/preview/internal/verify"
 	"github.com/coder/preview/types"
 )
-
-// Test_VerifyPreview will fully evaluate with `terraform apply`
-// and verify the output of `preview` against the tfstate. This
-// is the e2e test for the preview package.
-func Test_VerifyPreview(t *testing.T) {
-	t.Parallel()
-
-	installCtx, cancel := context.WithCancel(context.Background())
-
-	versions := verify.TerraformTestVersions(installCtx)
-	tfexecs := verify.InstallTerraforms(installCtx, t, versions...)
-	cancel()
-
-	dirFs := os.DirFS("testdata")
-	entries, err := fs.ReadDir(dirFs, ".")
-	require.NoError(t, err)
-
-	for _, entry := range entries {
-		entry := entry
-		if !entry.IsDir() {
-			t.Logf("skipping non directory file %q", entry.Name())
-			continue
-		}
-
-		entryFiles, err := fs.ReadDir(dirFs, filepath.Join(entry.Name()))
-		require.NoError(t, err, "reading test data dir")
-		if !slices.ContainsFunc(entryFiles, func(entry fs.DirEntry) bool {
-			return filepath.Ext(entry.Name()) == ".tf"
-		}) {
-			t.Logf("skipping test data dir %q, no .tf files", entry.Name())
-			continue
-		}
-
-		if slices.ContainsFunc(entryFiles, func(entry fs.DirEntry) bool {
-			return entry.Name() == "skip"
-		}) {
-			t.Logf("skipping test data dir %q, skip file found", entry.Name())
-			continue
-		}
-
-		name := entry.Name()
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			entryWrkPath := t.TempDir()
-
-			for _, tfexec := range tfexecs {
-				tfexec := tfexec
-
-				t.Run(tfexec.Version, func(t *testing.T) {
-					wp := filepath.Join(entryWrkPath, tfexec.Version)
-					err := os.MkdirAll(wp, 0755)
-					require.NoError(t, err, "creating working dir")
-
-					t.Logf("working dir %q", wp)
-
-					subFS, err := fs.Sub(dirFs, entry.Name())
-					require.NoError(t, err, "creating sub fs")
-
-					err = verify.CopyTFFS(wp, subFS)
-					require.NoError(t, err, "copying test data to working dir")
-
-					exe, err := tfexec.WorkingDir(wp)
-					require.NoError(t, err, "creating working executable")
-
-					ctx, cancel := context.WithTimeout(context.Background(), time.Minute*2)
-					defer cancel()
-					err = exe.Init(ctx)
-					require.NoError(t, err, "terraform init")
-
-					planOutFile := "tfplan"
-					planOutPath := filepath.Join(wp, planOutFile)
-					_, err = exe.Plan(ctx, planOutPath)
-					require.NoError(t, err, "terraform plan")
-
-					plan, err := exe.ShowPlan(ctx, planOutPath)
-					require.NoError(t, err, "terraform show plan")
-
-					pd, err := json.Marshal(plan)
-					require.NoError(t, err, "marshalling plan")
-
-					err = os.WriteFile(filepath.Join(wp, "plan.json"), pd, 0644)
-					require.NoError(t, err, "writing plan.json")
-
-					_, err = exe.Apply(ctx)
-					require.NoError(t, err, "terraform apply")
-
-					state, err := exe.Show(ctx)
-					require.NoError(t, err, "terraform show")
-
-					output, diags := preview.Preview(context.Background(),
-						preview.Input{
-							PlanJSONPath:    "plan.json",
-							ParameterValues: map[string]types.ParameterValue{},
-						},
-						os.DirFS(wp))
-					if diags.HasErrors() {
-						t.Logf("diags: %s", diags)
-					}
-					require.False(t, diags.HasErrors(), "preview errors")
-
-					if state.Values == nil {
-						t.Fatalf("state values are nil")
-					}
-					verify.Compare(t, output, state.Values.RootModule)
-				})
-			}
-		})
-	}
-}
 
 func TestFoo(t *testing.T) {
 	ty, err := gocty.ImpliedType([]any{1, 2, 3})
