+++ testdata/nulldefault/skipe2e
@@ -1,0 +1,1 @@
+No terraform to run.
